Enhancement to add support for larger fonts.
This change is backward compatible with the 1.0.11 version.

I recently have constructed additional oled fonts from a site I found that had character sizes > 8 pixels.  These fonts are useful for displays that are larger than 128x32 such as 128x64 or 128x128.  I plan to add these fonts in the very near future to npm but before I do, I need a module that will support them.

The current oled font character structure is such that each element in the array is a represents a column and bit values of the byte represent rows in the column.  The current code supports font with character widths larger than 8 pixels but it does not support fonts with character heights larger than 8 pixels.

This update includes a minor fix to word wrap (should no wrap the first word) and the changes to support fonts with character height larger than 8 pixels.
Version number bumped to 1.0.12.